backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
async-timeout==5.0.1
bidict==0.23.1
click==8.1.7
dnspython==2.7.0
fastapi==0.108.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mongomock==4.3.0
mongomock-motor==0.0.34
motor==3.6.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pymongo==4.9.2
pytest==8.3.4
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
redis==5.2.1
sentinels==1.0.0
setuptools==75.6.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.32.0.post1
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.32.1
websockets==14.1
wsproto==1.2.0
